Updated OpenRA to 20180218
Summary:
**Changelog:**

The main focus of this development cycle was modernizing core parts of the engine and game code to support the work-in-progress Tiberian Sun mod and other community projects. Thanks to our contributors, and to everybody who tested the playtest and reported bugs! Please note that changes to the way hotkeys are managed mean you will need to redefine any custom hotkeys in the Input settings.

Overhaul of the multiplayer UI added long-requested features:

* The server browser shows the name/faction/team of connected players
* The server browser is now also available as a tab in the server lobby
* The main menu and server browser will notify you if a new release is available
* The Global Chat has been removed from the server browser and lobby
* A chat dialog has been added to the in-game menu
* Server player list
* The new server player list is one of several improvements to the multiplayer UI.

Of course, we realize that the vast majority of OpenRA players do not play online. We haven't forgotten about you either! Improvements for the singleplayer experience include:

* The full Harkonnen campaign in D2K
* Refinements and improvements to several campaign missions in Red Alert and Tiberian Dawn
* A fix for the AI cheating when it uses superweapons and support powers
* Several other changes that should improve the AI behaviour and performance.
* Improvements to the AI support power targeting mean that this will be a rarer sight.

Other notable changes include:

* New lobby options to disable bounty, build radius, and re-packing Construction Yards
* Units will no longer target enemies revealed by the Red Alert GPS Satellite
* A collection of community-driven balance changes for Red Alert and Tiberian Dawn
* Fixes for many graphical glitches and polish issues
* Fixes for several platform integration issues on Windows and macOS
* Further improvements to memory usage and performance
* Expanded game options
* The Map Options tab includes several new settings, and can be customized by scripted maps.

Test Plan: Opened OpenRA-Dune 2000 and started a new game. Everything worked as expected.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.solus-project.com/D2326